+++ src/main/java/com/blackrook/small/SmallConstants.java
@@ -19,6 +19,8 @@ public interface SmallConstants
 	String SMALL_APPLICATION_CONFIGURATION_ATTRIBUTE = "small.application.configuration";
 	/** The attribute name for the path remainder that gets set on the request context on each request (default paths only). */
 	String SMALL_REQUEST_ATTRIBUTE_PATH_REMAINDER = "small.request.path.remainder";
+	/** The attribute name for the SmallResponse that is passed around between filters and calls (may be null). */
+	String SMALL_REQUEST_ATTRIBUTE_RESPONSE_OBJECT = "small.request.response.object";
 	/** The attribute name for the part list on multipart requests. */
 	String SMALL_REQUEST_ATTRIBUTE_MULTIPART_LIST = "small.request.part.list";
 }